add demo scripts
