+++ site/package.json
@@ -8,7 +8,9 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "format": "prettier --w -l ."
+    "format": "prettier --w -l .",
+    "demo:generate": "tsx --env-file=.env.local scripts/demo/generate.ts",
+    "demo:clean": "tsx --env-file=.env.local scripts/demo/clean.ts"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.4",
@@ -54,6 +56,7 @@
     "prettier-plugin-tailwindcss": "^0.5.11",
     "prisma": "^5.8.1",
     "tailwindcss": "^3.3.0",
+    "tsx": "^4.7.1",
     "typescript": "^5"
   }
 }
