+++ site/scripts/demo/clean.ts
@@ -1,0 +1,7 @@
+import prisma from "@/lib/prisma/client";
+
+prisma.listing
+  .deleteMany()
+  .then((res) =>
+    console.info(`Deleted listings: ${JSON.stringify(res, null, 2)}`),
+  );
